+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/Keytar/Keytar.cs
@@ -69,7 +69,7 @@ namespace PlasticBand.Devices
         /// <summary>
         /// The Select button on the keytar.
         /// </summary>
-        [InputControl(displayName = "Back")]
+        [InputControl(displayName = "Select")]
         public ButtonControl selectButton { get; private set; }
 
         /// <summary>