+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/ProGuitar/ProGuitar.cs
@@ -74,7 +74,7 @@ namespace PlasticBand.Devices
         /// <summary>
         /// The Select button on the guitar.
         /// </summary>
-        [InputControl(displayName = "Back")]
+        [InputControl(displayName = "Select")]
         public ButtonControl selectButton { get; private set; }
 
         /// <summary>