+++ Packages/com.thenathannator.plasticband/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Dates are relative to UTC.
 
+## Unreleased
+
+## Fixed
+
+- Fixed some devices incorrectly using "Back" for their `selectButton` display name instead of "Select".
+
 ## [0.2.1] - 2023/14/04
 
 ### Fixed